services/auth-service/requirements.txt
alabaster==1.0.0
autodoc==0.5.0
babel==2.18.0
beautifulsoup4==4.14.3
black==26.3.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
coverage==7.13.5
decorator==5.2.1
docutils==0.22.4
flake8==7.3.0
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
idna==3.11
imagesize==2.0.0
iniconfig==2.3.0
isort==8.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
legacy-cgi==2.6.4
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.2
pytest-cov==7.1.0
python-dotenv==1.2.2
pytokens==0.4.1
requests==2.33.1
roman-numerals==4.1.0
snowballstemmer==3.0.1
soupsieve==2.8.3
Sphinx==9.1.0
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
waitress==3.0.2
WebOb==1.8.9
WebTest==3.0.7
Werkzeug==3.1.6
